Split Action and Pref annotations into own packages
Make them more discoverable. Also remove `@ActionDescription`, since `@ActionConfig` should be used instead.
