+++ qupath-extension-processing/src/main/java/qupath/process/gui/WandToolEventHandler.java
@@ -60,11 +60,11 @@ import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.ObjectProperty;
 import javafx.scene.input.MouseEvent;
 import qupath.lib.gui.QuPathGUI;
-import qupath.lib.gui.panes.PreferencePane.BooleanPref;
-import qupath.lib.gui.panes.PreferencePane.DoublePref;
-import qupath.lib.gui.panes.PreferencePane.Pref;
-import qupath.lib.gui.panes.PreferencePane.PrefCategory;
 import qupath.lib.gui.prefs.PathPrefs;
+import qupath.lib.gui.prefs.annotations.BooleanPref;
+import qupath.lib.gui.prefs.annotations.DoublePref;
+import qupath.lib.gui.prefs.annotations.Pref;
+import qupath.lib.gui.prefs.annotations.PrefCategory;
 import qupath.lib.gui.viewer.QuPathViewer;
 import qupath.lib.gui.viewer.overlays.HierarchyOverlay;
 import qupath.lib.gui.viewer.overlays.PathOverlay;
