+++ qupath-extension-svg/src/main/java/qupath/lib/extension/svg/SvgExtension.java
@@ -28,8 +28,8 @@ import qupath.lib.extension.svg.SvgExportCommand.SvgExportType;
 import qupath.lib.gui.QuPathGUI;
 import qupath.lib.gui.QuPathResources;
 import qupath.lib.gui.actions.ActionTools;
-import qupath.lib.gui.actions.ActionTools.ActionConfig;
-import qupath.lib.gui.actions.ActionTools.ActionMenu;
+import qupath.lib.gui.actions.annotations.ActionConfig;
+import qupath.lib.gui.actions.annotations.ActionMenu;
 import qupath.lib.gui.extensions.QuPathExtension;
 
 /**
