+++ qupath-gui-fx/src/main/java/qupath/lib/gui/actions/OverlayActions.java
@@ -25,9 +25,9 @@ package qupath.lib.gui.actions;
 import org.controlsfx.control.action.Action;
 
 import qupath.lib.gui.SelectableItem;
-import qupath.lib.gui.actions.ActionTools.ActionAccelerator;
-import qupath.lib.gui.actions.ActionTools.ActionConfig;
-import qupath.lib.gui.actions.ActionTools.ActionIcon;
+import qupath.lib.gui.actions.annotations.ActionAccelerator;
+import qupath.lib.gui.actions.annotations.ActionConfig;
+import qupath.lib.gui.actions.annotations.ActionIcon;
 import qupath.lib.gui.commands.Commands;
 import qupath.lib.gui.tools.IconFactory.PathIcons;
 import qupath.lib.gui.viewer.OverlayOptions;
