+++ qupath-gui-fx/src/main/java/qupath/lib/gui/actions/annotations/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Annotations that can be used to configure actions.
+ */
+package qupath.lib.gui.actions.annotations;
