+++ qupath-gui-fx/src/main/java/qupath/lib/gui/actions/menus/EditMenuActions.java
@@ -15,9 +15,9 @@ import qupath.lib.gui.QuPathResources;
 import qupath.lib.gui.UndoRedoManager;
 import qupath.lib.gui.actions.ActionTools;
 import qupath.lib.gui.actions.CommonActions;
-import qupath.lib.gui.actions.ActionTools.ActionAccelerator;
-import qupath.lib.gui.actions.ActionTools.ActionConfig;
-import qupath.lib.gui.actions.ActionTools.ActionMenu;
+import qupath.lib.gui.actions.annotations.ActionAccelerator;
+import qupath.lib.gui.actions.annotations.ActionConfig;
+import qupath.lib.gui.actions.annotations.ActionMenu;
 import qupath.lib.gui.commands.Commands;
 import qupath.lib.gui.tools.GuiTools;
