+++ qupath-gui-fx/src/main/java/qupath/lib/gui/actions/menus/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Classes to store actions that correspond to QuPath's main menus.
+ * <p>
+ * Representing the menus in this way is intended to make it easier to find core functionality.
+ */
+package qupath.lib.gui.actions.menus;
