+++ qupath-gui-fx/src/main/java/qupath/lib/gui/actions/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Classes to manage Actions.
+ * <p>
+ * This encapsulate QuPath's functionality in a way that is separate from specific UI elements.
+ * It's possible to then generate (for example) one or more menu items and buttons for each action, 
+ * while binding the text, graphic and tooltip to the data stored in the action.
+ */
+package qupath.lib.gui.actions;
