+++ qupath-gui-fx/src/main/java/qupath/lib/gui/prefs/annotations/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Annotations that can be used to define QuPath preferences.
+ */
+package qupath.lib.gui.prefs.annotations;
