+++ qupath-gui-fx/src/main/resources/qupath/lib/gui/localization/qupath-strings.properties
@@ -153,7 +153,7 @@ ViewerActions.matchResolutions.description = Adjust zoom factors to match the re
 ViewerActions.zoomToFit = Zoom to fit
 ViewerActions.zoomToFit.description = Adjust zoom for all images to fit the entire image in the viewer
 
-# Default commands for a QuPath instance
+# Shared commands for a QuPath instance
 CommonActions.objectDescriptions = Show object descriptions
 CommonActions.objectDescriptions.description = Show descriptions for the currently-selected object, where available.\nDescriptions can be any plain text, markdown or html added as the 'description' property to an object (currently, only annotations are supported).
 CommonActions.measureTMA = Show TMA measurements
